tests/example_project/views/authentication.py
from django.contrib.auth.models import User
from rest_framework import serializers
from rest_framework.decorators import api_view
from rest_framework.request import Request
from rest_framework.response import Response


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = (
            'username',
            'first_name',
            'last_name',
            'email',
        )


@api_view()
def user_info(request: Request) -> Response:
    serializer = UserSerializer(request.user)
    return Response(serializer.data)
