pyproject.toml
[tool.poetry]
name = "pytest-drf"
version = "0.1.0"
description = "A Django REST framework plugin for pytest."
authors = ["Zach \"theY4Kman\" Kanzler <they4kman@gmail.com>"]
keywords = ["pytest", "django", "djangorestframework"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/theY4Kman/pytest-drf"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Framework :: Django",
  "Framework :: Pytest",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Testing",
]

[tool.poetry.dependencies]
python = ">=3.6,<4.0"

djangorestframework = ">3"
inflection = "^0.3.1"
pytest = ">=3.6"
pytest-assert-utils = "^0.1.0"
pytest-common-subject = "^1.0"
pytest-lambda = "^1.1"

[tool.poetry.dev-dependencies]
pytest-django = "^3.9.0"
pytest-camel-collect = "^1.0.1"
django = "^3"

[tool.poetry.plugins."pytest11"]
drf = "pytest_drf.plugin"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
